+++ petting_area/__init__.py
@@ -1,0 +1,5 @@
+from .donkey import Donkey
+from .goat import Goat
+from .lamb import Lamb
+from .llama import Llama
+from .pig import Pig
