+++ pond/__init__.py
@@ -1,0 +1,5 @@
+from .beaver import Beaver
+from .catfish import Catfish
+from .duck import Duck
+from .goldfish import Goldfish
+from .turtle import Turtle
